seems to be working

+++ modexp.cpp
@@ -5,14 +5,14 @@
 
 using namespace std;
 
-int modexp(int a, int b, int c);
+long long modexp(long long a, long long b, long long c);
 
 void get_input(int *a, int *b, int *c);
 
 int main() {
-	int a, b, n;
+	long long a, b, n;
 	get_input(&a, &b, &n);
-	int answer = modexp(a, b, n);
+	long long answer = modexp(a, b, n);
 
 	cout << a << "^" << b << " mod "<< n << " = ";
 	cout << answer << endl; 
@@ -20,8 +20,8 @@ int main() {
 	return 0;
 }
 
-void get_input(int *a, int *b, int *n) {
-	int arr[3];
+void get_input(long long *a, long long *b, long long *n) {
+	long long arr[3];
 	string args;
 
 	cout << "Enter the values of a b and n to compute a^b mod n: ";
@@ -33,7 +33,7 @@ void get_input(int *a, int *b, int *n) {
 
 	while (ssin.good() && i < 3) {
 		ssin >> temp;
-		arr[i] = stoi(temp);
+		arr[i] = stol(temp);
 		i ++;
 	}
 
@@ -43,8 +43,8 @@ void get_input(int *a, int *b, int *n) {
 
 }
 
-int modexp(int a, int b, int n) {
-	int pow = 1;
+long long modexp(long long a, long long b, long long n) {
+	long long pow = 1;
 	for (int i = 0; i < b; i++)
 		pow *= a;
 	return pow % n;
